requirements.txt
hyperliquid-python-sdk>=0.9.0
python-dotenv>=1.0.0
eth-account>=0.11.0
certifi>=2024.0.0
flask>=3.0.0
